move node editor to shared folder
