+++ src/backoffice/editors/shared/node/editor-node.element.ts
@@ -2,20 +2,20 @@ import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { css, html, LitElement } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { UUIInputElement, UUIInputEvent } from '@umbraco-ui/uui';
-import { UmbContextConsumerMixin, UmbContextProviderMixin } from '../../../core/context';
-import { UmbNodeStore } from '../../../core/stores/node.store';
+import { UmbContextConsumerMixin, UmbContextProviderMixin } from '../../../../core/context';
+import { UmbNodeStore } from '../../../../core/stores/node.store';
 import { distinctUntilChanged, Subscription } from 'rxjs';
-import { NodeEntity } from '../../../mocks/data/node.data';
-import type { UmbNotificationService } from '../../../core/services/notification';
+import { NodeEntity } from '../../../../mocks/data/node.data';
+import type { UmbNotificationService } from '../../../../core/services/notification';
 import { UmbNodeContext } from './node.context';
 
-import '../shared/editor-entity/editor-entity.element';
+import '../../shared/editor-entity/editor-entity.element';
 
 // Lazy load
 // TODO: Make this dynamic, use load-extensions method to loop over extensions for this node.
 import './views/edit/editor-view-node-edit.element';
 import './views/info/editor-view-node-info.element';
-import { UmbNotificationDefaultData } from '../../../core/services/notification/layouts/default';
+import { UmbNotificationDefaultData } from '../../../../core/services/notification/layouts/default';
 
 @customElement('umb-editor-node')
 export class UmbEditorNodeElement extends UmbContextProviderMixin(UmbContextConsumerMixin(LitElement)) {
